+++ src/device-private.h
@@ -33,7 +33,7 @@
 extern "C" {
 #endif /* __cplusplus */
 
-#define EVENT_PROGRESS_INITIALIZER {0, UINT_MAX}
+#define EVENT_PROGRESS_INITIALIZER {0, SIZE_MAX}
 
 struct dc_device_t;
 struct dc_device_vtable_t;
@@ -92,7 +92,7 @@ int
 device_is_cancelled (dc_device_t *device);
 
 dc_status_t
-device_dump_read (dc_device_t *device, unsigned int address, unsigned char data[], unsigned int size, unsigned int blocksize);
+device_dump_read (dc_device_t *device, unsigned int address, unsigned char data[], size_t size, unsigned int blocksize);
 
 #ifdef __cplusplus
 }
